Better reporting of timeouts
Update the check_* functions to return a status of the result ("correct",
"timeout", or "incorrect"), and make the plugins/tools print out a message
for timeouts too.

The most serious status code is returned. For example, doing a refinement
check where the "retval" check times out, and the "UB" check detects
that the transformation is incorrect will return "incorrect".