Eliminate some redundant range constraints
The SMT solver get confused, and becomes much slower, when we have
both a mask and a range for constraining a value. For example, one
test I analyzed with a loop

  for (i = 0; i < n; i++)

took 95 seconds to analyze when we constrained i as both

  0 <= i && i <= 0x7fffffff

and

  (i & 0x80000000) == 0

But it took only 5 seconds to analyze when we only used

  0 <= i && i <= 0x7fffffff

We therefore should avoid generating the mask check when it does not
constrain the value more than what the range already does.

This patch is a stop gap to take care of some trivial cases.