Speed up checking
I noticed check_refine_z3 spent a lot of time destroying the
context -- it often spent more time in the destructor than
time spent checking the IR! The reason turned out to be
that it is slow to free "leaked" AST nodes (where the
reference count is not 0).

But we are using the Z3 C++ API that is supposed to handle the
reference counting, so there should not be any leaked nodes...
But there is an issue in the API when a z3::expr is used both
in lhs and rhs, such as
   x = x + y;
This leaks the original x, while doing this as
   z3::expr t = x + y;
   x = t;
works fine. And this kind of construct were used, a lot,
in stmtgcc when updating the arrays for z3::store, and
when updating the conditions for UB handling and control
flow.

This patch updates all of those to not reuse variables.