Fix fmin/fmax issue with -0.0
Floating point treats 0.0 and -0.0 as equal, so fmax(0.0, -0.0) may
return either 0.0 or -0.0. We returned the 2nd argument when both
arguments were equal, but that made smtgcc-tv report that the function
below is miscompiled

  void
  f1 (double *res, double x, double y)
  {
    res[0] = __builtin_fmax (x, y);
    res[1] = __builtin_fmax (y, x);
  }

because fwprop swaps the arguments for one of the fmax so the two
calls can be CSE:ed, and we now fail memory checking for the case
where x is 0.0 and y is -0.0.

This patch fix this by always returning 0.0 for both fmax(0.0, -0.0)
and fmax(-0.0, 0.0), and similarly returning -0.0 for fmin(0.0, -0.0)
and fmin(-0.0, 0.0).