Remove redundant solver constraint
The check that tgt does not have UB that is not in src used two
equivalent constraints
  solver.add(tgt_ub_expr);
and
  solver.add(src_ub_expr != tgt_ub_expr);
because that was slightly faster for the limited test cases I could
run at that time. But this seems to confuse the Z3 heuristics when
analyzing complex functions, so it faster to only keep one of them.