Improve CLZ and CTZ internal functions
The internal functions CLZ and CTZ may return an undefined value
which we implement as a SYMBOLIC value. But we used a different
symbolic value for each pass, which mean that SMTGCC reports
a miscompilation if, for example, that value is stored in global
memory, as the memory will be different.

This is fixed by handling the SYMBOLIC instruction in the same way
as the PARAM instruction: adding an index argument and using the
same index for the src and tgt functions.