Move part of IR -> SMT conversion to a separate pass
Many GCC optimizations only makes minor changes to the IR, so most of
the code is identical for src and tgt. But Z3 sometimes has problems
dealing with this -- I have seen cases where, for example, the UB
state for src and tgt are identical, but Z3 times out when trying to
verify that tgt does not have more UB than what src has.

This patch is the first step in improving this. We now eliminate the
control flow at the IR level, and CSE between src and tgt. This gives
Z3 simpler code to work with, and is often faster to check.