Fix more z3::expr leaks
The Z3 reference counting apparently has problems when updating
an expr with a non-trivial calculation (even when the target expr
is not used in the calculation). So we must create a new expr
for the result of the calculation, and then copy it to the real
destination. For example, we must change

  is_more_undef = (src_mask & tgt_undef) != 0;

to

  z3::expr new_is_more_undef = (src_mask & tgt_undef) != 0;
  is_more_undef = new_is_more_undef;