Improve simplification of wide Boolean
GCC use a type "wide Boolean", such as <signed-boolean:32>, for
Boolean in vectors. I originally emitted them as 1-bit values, and
extended them when needed, but that caused all kinds of problems,
so I now always use the size GCC expects. But the values are usually
created from an 1-bit value (such as Op::EQ), and used where a
Boolean is needed (such as the condition for Op::ITE), so we extend
and truncate them.

Most of these extensions/truncation are just right after each other

  %300 = sext %206, %18
  %301 = extract %300, %6, %6

and they are eliminated by simplify(). But some common cases were
not simplified, such as

  %343 = sext %206, %18
  %344 = sext %211, %18
  %345 = and %343, %344
  %351 = extract %345, %6, %6

that could have been done as just

  %351 = and %206, %211

and similar for Op::OR/Op::XOR/Op::NOT. There is also a missing case
for Op::ITE where we failed to change
  ite a, -1, 0 -> sext a

Fixing this helps Z3 a lot for vectorized code. For example, it
eliminates about 30% of the time for checking gcc.dg/vect/vect-bool-cmp.c