Improve UB generation
For basic blocks with a common condition in src and tgt, we split
the list of UB into three sets:
 * the ones identical in both
 * the set unique to src
 * the set unique to tgt
The common UB is generated first (so it is CSE'd between src and tgt),
followed by the other two sets.

This greatly assists the SMT solver when checking that tgt does not
have more UB than src, especially since it only needs to examine the
UB that are unique to tgt.