Remove identical()
The identical() function was implemented as a stop gap for not
running the solver when nothing was changed by the GCC pass
(because destroying the solver context was very slow due to the
leaked expr issue).

We now have a much better solution in the new Converter (even
though it is not as important after implementing the workaround
for the Z3 reference counting issues).