Improve store-to-load forwarding
Store-to-load forwarding can be done much more effectively
in the converter, as the memory state there is explicitly
represented in the form of an immutable array.

We also move the load instructions as early as possible in
the chain of arrays. This increases the likelihood that load
instructions can be CSE'd.