Remove check for overly large programs
The checks for MAX_BBS and MAX_NOF_INSTS were previously performed
after the IR had been created. This approach made sense a while
back -- we aimed to stress the GIMPLE->IR converter, and some of
the later IR passes were very expensive.

However, the later IR passes have been optimized -- it is now the
GIMPLE->IR conversion that is the expensive part! Therefore, we
might as well run the full pass list. Most of the tests will time
out in the SMT solver, but since it results in just one timeout,
it does not significantly affect the total time when running the
full test suite.