Initial implementation of pointer provenance
This assigns a "provenance" to each pointer when it is derived from
an expression involving VAR_DECL. Pointer arithmetic retains the
provenance from the original pointer, while casts from integer to
pointer "invent" a provenance based on the address.

The provenance corresponds to the mem_id of the address. Each memory
access is then checked to ensure that the accessed address has the
same mem_id as the provenance. If not, the access invokes undefined
behavior.

Note: We are not yet propagating provenance through store/load.