Workaround for SMT solver NaN issue
SMT solvers has only one NaN value, so NEGATE_EXPR of NaN does not
change the value. This leads to incorrect reports of miscompilations
for transformations like -ABS_EXPR(x) -> .COPYSIGN(x, -1.0), either
because the returned/memory values differ or because copysign has
introduceed a non-canonical NaN (as our implementation of copysign
works on the bitvector instead of working on a floating point value).

For now, treat negation of NaN as UB to avoid these false reports of
miscompilation.