Emit vector phis as one phi per element
We implement vector operations by splitting the vectors, performing
scalar instructions, and then concatenating the scalar results into
a new vector. The intention is for the concatenations and extractions
to be eliminated by simplify_inst. However, this approach does not work
for phi-nodes, resulting in many unnecessary concatenations before the
phi nodes and extractions after them -- especially in unrolled loops.

To resolve this, we are now emitting vector phi-nodes on an
element-by-element basis.