Improve simplify_concat
Store-to-load forwarding of vectors often results in sequences like:

  %1137 = extract %976, %6, %4
  %1138 = extract %976, %8, %7
  %1139 = extract %976, %10, %9
  %1140 = extract %976, %12, %11
  %1161 = concat %1037, %664
  %1162 = concat %1038, %1161
  %1163 = concat %1039, %1162
  %1164 = concat %1040, %1163
  %1141 = extract %988, %6, %4
  %1142 = extract %988, %8, %7
  %1143 = extract %988, %10, %9
  %1144 = extract %988, %12, %11
  %1165 = concat %1041, %1164
  %1166 = concat %1042, %1165
  %1167 = concat %1043, %1166
  %1168 = concat %1044, %1167
  ...

Previously, simplify_concat could eliminate a concat of extract when
the extractions were consecutive from one object. However, it didn't
address the case where each vector scalar value is a distinct object
and elements of a second object are concatenated to the first.

Now, we optimize this pattern to:

  %1164 = concat %976, %664
  %1168 = concat %988, %1164