Improve sorting of BB preds and phi arguments
We need to sort the phi node arguments and BB predecessors in a
consistent way to prevent random differences between src and tgt
when we create ITEs for phi nodes. We basically sort them in
reverse post order, but we treat basic blocks having exactly one
predecessor and successor as having the same position as its
successor.

This approach generates the same ITEs for cases such as the GCC
pre pass eliminating BB3 as shown below (where BB2 in the original
came before BB3 in RPO)

       BB1                          BB1
       / \                           |\
      /   \                          | \
    BB3   BB2          ===>          | BB2
      \   / \                        | / \
       \ /   \                       |/   \
       BB4   ...                    BB4   ...
        |                            |
       ...                          ...

which would differ if we used the reverse post order directly.