smtgcc-opt: make -c work for -simplify_inst
The converter runs inst_combine when creating the final function to
send to the SMT solver. Consequently,
  smtgcc-opt -simplify_inst -c test.ir
does not perform any checks, as the source version used as the
reference is transformed in the same way as tgt during conversion.

This patch allows for the disabling of inst_combine in the converter
and disables it specifically for smtgcc-opt.