Tweak mem_flag handling
mem_flag tracking was made more precise in commit 8d36a5b. However,
this fails when GCC optimizes individual pointer assignments like:

  f.p = &y;
  f.q = &x;

into a vector store:

  _22 = VIEW_CONVERT_EXPR<long unsigned int>(&x);
  _30 = VIEW_CONVERT_EXPR<long unsigned int>(&y);
  _28 = {_30, _22};

In such cases, mem_flag is not set to 1. To resolve this, we revert a
portion of 8d36a5b to ensure mem_flag is set to 1 in these scenarios.