Make BIT_FIELD_REF more efficient
Consider the function f6 from
testsuite/c-c++-common/torture/builtin-convertvector-1.c:

  void
  f6 (v256df *x, v256di *y)
  {
    *y = __builtin_convertvector (*x, v256di);
  }

The original GIMPLE is emitted as

  _2 = *x_4(D);
  _7 = BIT_FIELD_REF <_2, 64, 0>;
  _9 = BIT_FIELD_REF <_2, 64, 64>;
  _11 = BIT_FIELD_REF <_2, 64, 128>;
  ...

and this is modified by forwprop4 to

  _7 = BIT_FIELD_REF <*x_4(D), 64, 0>;
  _9 = BIT_FIELD_REF <*x_4(D), 64, 64>;
  _11 = BIT_FIELD_REF <*x_4(D), 64, 128>;
  ...

This is functionally equivalent to the original, but we generated
IR that for each BIT_FIELD_REF loaded a full vector and then
extracted the element, which produces lots of useless instructions
that later passes needed to eliminate.

Now, we've modified the generation of each BIT_FIELD_REF to only
load the bytes necessary.