Fix CFN_LOOP_VECTORIZED false positives
We emit CFN_LOOP_VECTORIZED as a symbolic value, which ensures
that we verify tgt is a refinement of src regardless of its value.
However, this approach can lead to false positives in cases where
gcc decides not to vectorize the loop. The issue arises because
the 'true' branch of CFN_LOOP_VECTORIZED may have changed some
integer variables to unsigned to prevent overflow. When
CFN_LOOP_VECTORIZED is eliminated, we then detect a possible
overflow when this is compared to tgt, which now only contains
the original, overflowing calculations.

We address this by setting CFN_LOOP_VECTORIZED to `false` in src
if tgt does not include any CFN_LOOP_VECTORIZED.