remove_unused_memory: remove worklist.reserve
Enabling support for large global variables made worklist.reserve
reserve gigabytes for some test cases.

The worklist.reserve had some relevance when it was added -- we did
not simplify address calculations, so the code setting local values
to uninitialized had long chains of ptr+1+1+1+1 for the full memory
size. But the work list now typically just end up with two elements,
so we may as well just remove the reservation.