Always constrain Booleans
We must constrain Booleans because the vectorizer may generate code
that assumes Booleans have the value 0 or 1.

This can be seen when compiling the following function for X86_64
with -O3 -march=x86-64-v4

bool vect_a[N];
unsigned test4(bool x)
{
 unsigned ret = 0;
 for (int i = 0; i < N; i++)
 {
   if (vect_a[i] == x)
     return 1;
   vect_a[i] = x;
 }
 return ret;
}

The vectorized code reads vect_a as an array of unsigned char and
compares the elements with (unsigned char)x. This transformation
is not valid if the bytes may have values different from 0 or 1.