Disable checking of CFN_LOOP_VECTORIZED
Introduce a global variable to control whether we should verify
that `tgt` is a refinement of `src` for both possible values of
CFN_LOOP_VECTORIZED, defaulting to false.

Setting this to true makes us check that tgt is a refinement of
src for both possible values of CFN_LOOP_VECTORIZED. This slows
down the checking and reports issues for cases that are arguably
correct (such as in gcc.c-torture/execute/pr94734.c where the
true case has been changed from
  if (x == (i & 0x25))
    arr[y] = i;
to
  _25 = &arr[y_12(D)];
  .MASK_STORE (_25, 32B, _2, i_17);
which may result in invalid indexing for cases where
  x == (i & 0x25)
is false, but would not be a problem after vectorization.