More aggressive "always UB" optimization
We cannot remove used instructions if the function has loops,
since this may change the control flow so that there is no
path to the exit block. However, we may still remove store
instructions, etc.

Also, ensure that the "ub 1" instruction is the first in the
BB so that UB propagation triggers for preceding BBs.