Check for extra UB before retval and memory
We checked for incorrectly introduced UB last, as an extra UB that
does not change the result (such as a signed integer overflow
introduced during vectorization) generally has a lower priority
to fix compared to issues that do change the result. However,
recent changes to DCE may affect the solver results for all
incorrectly introduced UB, even when the code generated by GCC
does not. This leads to confusing reports of a retval/memory
difference when the issue is actually an extra UB.

Fix this by performing the UB check first.