Check if the GCC pass has modified the IR
We save execution time in check_refine by not running the SMT
solver if the src and tgt functions are identical. However, we
are now translating src and tgt slightly differently for some
cases (constraints, etc.), so we often end up with differences
in the IR even when the GIMPLE is identical...

This commit adds an identity check in smtgcc-tv that avoids
running the solver when the GIMPLE is identical.