Change how uninit is handled
Using uninitialized values is now UB for instructions that are UB
for certain values (such as a signed integer PLUS_EXPR, which is
UB if the calculation overflows).

The rationale behind the change is that uninitialized values can
be any value, potentially invoking the UB. But the implementation
needs some improvements. For example, we know that "uninit + 0"
cannot overflow, so this should arguably be allowed.