Simplify load/store of undef byte
We had a special case in process_load/process_store to skip undef
bytes. This is fine as the value will never be used anyway (as it
will fail UB or undef checks), and it gives us slightly simpler IR
when reading/writing padding. But it has the side effect of preventing
CSE between src and tgt for some optimizations (especially when
comparing GIMPLE with assembly), making the IR much slower to check.