Remove the flattening and/or elimination
Remove the optimizations:
 * and x, (and y, (and x, z)) -> and y, (and x, z)
 * or x, (or y, (or x, z)) -> or y, (or x, z)

These do not make sense before CSE, and the implementation is
quadratic in the number of chained instructions. The implementation
was also too limited anyway, as it only checked if a single instruction
existed in the tree of and/or instead of eliminating all redundant
instructions when merging two trees.

A better implementation will be added in check.cpp at a later time.