riscv: more permissive load UB check
Relax the out-of-bounds check to allow reading a few bytes
after a memory object, as the RISC-V backend sometimes uses
a wider load than the object size. For example, consider the
variable `c` defined as follows:

     .globl  c
     .align  2
     .type   c, @object
     .size   c, 3
  c:
     .zero   3

This variable may be read using a four-byte load, which is
acceptable because the alignment ensures at least one byte
is available beyond the object.