Improve IR canonical representation
RISC-V assembly likes to mix shifts with a constant offset and
arithmetic instructions to implement arithmetic on values less
than the register width, and we were not that good at simplifying
that.

Another related problem has to do with UB checks for the GCC ranges.
We often have code of the form "((x * 4) & 3) == 0" that we were not
good at eliminating.

This change is a first step in fixing this. We now canonicalize most
instructions that introduce 0-valued bits at the top and the bottom
of a value as "concat x, 0," which makes it much easier to track
both which bits are used and which are zero.