Fix instruction order problem
We were building the "check" function by adding new instructions
at the end. The problem with this is that the CSE could add some
extra instructions (when changing a comparison to its canonical
version) which were placed last. This caused the instructions to
be in the wrong order if we tried to CSE an already created
instruction (which happens if simplify_inst has introduced new
instructions that we then CSE).

Fix this by ensuring the CSE code moves new instructions to the
correct place.