Improve IR canonical representation
Commit 43b54b2 changed the IR canonical representation to be explicit
with 0-valued bits by using "concat x, 0," etc. when possible. This
was a bad idea -- the Z3 heuristics did not like the new code, and
analysis often regressed in performance.

The reason for that representation was to make it trivial to remove
redundant UB checks for the GCC ranges, but that is now handled by
the VRP pass. We can therefore revert the harmful part of 43b54b2.