Rename "undef"
Our semantics for "undef" are that the bits have wobbly values. But
the term "undef" is confusing as it has a different meaning in C and
LLVM, among others. This is especially confusing in gimple_conv.cpp,
where we mix the semantics of C, GIMPLE, and our IR.

We therefore need a different name, preferably one that does not have
a special meaning in C. I have chosen "indefinite." This commit
updates all variable names and comments to use the new name.