Make indef handling more precise
We reported that the veclower pass introduced new UB because we were
not sufficiently precise in our indef handling.

The issue occurs for vectors of wide booleans:

  vector(2) <signed-boolean:64> _1;

where the original GIMPLE does a comparison:

  _1 = v_2(D) > { 0, 0 }

This is fine if v_2 is indef -- _1 will in that case also be indef.
But veclower changes this to:

  _4 = BIT_FIELD_REF <v_2(D), 64, 0>;
  _5 = _4 > 0;
  _6 = (signed long) _5;
  _7 = -_6;
  _8 = (<signed-boolean:64>) _7;
  _9 = BIT_FIELD_REF <v_2(D), 64, 64>;
  _10 = _9 > 0;
  _11 = (signed long) _10;
  _12 = -_11;
  _13 = (<signed-boolean:64>) _12;
  _1 = {_8, _13};

which gives us new UB from _7 and _12.

We fix this by updating type conversion to mark the extended bits
as defined for zero extension or sign extension where the most
significant bit is defined. We also update NEGATE_EXPR and ABS_EXPR
to only report UB from wrapping if the value may be the minimal
integer when assigning concrete values to the indef bits.