Improve workaround for non-canonical NaN
Backends may choose to implement fabs and fneg as bit operations. Skip
checking if this would generate a non-canonical NaN.

Similar issues may occur with copysign and xorsign, where the Converter
had special cases to canonicalize NaNs that couldn't be applied when
parsing the generated code. Remove this special-casing and skip checking
if the result is a non-canonical NaN.