smtgcc-tv-backend: unroll one extra iteration
The backend may transform loops so that the assembly loop iterates
one extra iteration. We must therefore unroll loops in tgt one extra
iteration to prevent false reports that tgt is more UB than src.

One example of this can be seen for the loop:

  int foo (int n)
  {
    <bb 2> :
    if (n_2(D) > 44)
      goto <bb 3>;
    else
      goto <bb 4>;

    <bb 3> :
    # n_6 = PHI <n_4(3), n_2(D)(2)>
    n_4 = n_6 + -45;
    if (n_4 > 44)
      goto <bb 3>;
    else
      goto <bb 4>;

    <bb 4> :
    # n_7 = PHI <n_4(3), n_2(D)(2)>
    return n_7;
  }

This loop is generated in AArch64 assembly as

  foo:
          b       .L5
  .L3:
          sub     w0, w0, #45
  .L5:
          cmp     w0, 44
          bgt     .L3
          ret

where the exit check has been moved to the top of the next iteration.