Don't special-case floating-point in ITE
We attempted to generate efficient code by avoiding translation between
floating-point and bitvectors when performing ITE. But this approach can
give us an incorrect result when the code use a value both as floating-
point and as bits.

A simple failure case is

  int foo(int c, float a, float b) {
     float f = a + b;
     int i;
     memcpy(&i, &f, 4);
     return c ? i : -1;
  }

where we converted -1 to floating-point for the ITE, which is wrong
as it get translated to the solver's canonical NaN. More complex
cases can happen where, for example, a loaded value is used both as
a floating-point value and a bitvector. We must therefore disable
this special-casing for floating-point and always perform ITE as
bitvectors.