Improve handling of non-standard Boolean types
GCC has non-standard Boolean types, such as signed-boolean:8, that
have more than one bit. The handling in the generated IR was somewhat
inconsistent -- some usage kept the full size, while some treated
them as 1-bit values extended on demand.

This worked well enough, but there were cases where we could fail
to detect incorrect usage, such as a signed Boolean having a value
other than 0 or -1, or an unsigned Boolean having a value other
than 0 or 1.

This patch changes the implementation to always use the full width,
with checks that ensures that values are valid for the type.