Improve workaround for non-canonical NaN
The improved workaround for non-canonical NaN in 98e00ea simplified
the code a bit too much, which may, in some cases, lead to false
positives.

One example of the failure is:

  float foo (float x)
  {
    float _1;
    float _3;

    <bb 2> :
    _1 = __builtin_copysignf (1.0e+0, x_2(D));
    _3 = _1 * x_2(D);
    return _3;
  }

which the widening_mul pass optimizes to:

  float foo (float x)
  {
    float _3;

    <bb 2> :
    _3 = .XORSIGN (x_2(D), x_2(D));
    return _3;
  }

We report this as incorrect if x is NaN because the multiplication
in the original code returns a canonical NaN, while the optimized
code returns a non-canonical NaN.

This fix brings back the NaN canonicalization that was removed
in 98e00ea.