Reduce max_nof_inst
The original limit on the number of instructions was chosen to
prevent memory explosions when building the SMT and was set much
higher than what is reasonable for the SMT solver to check.
Reduce the value to skip more functions that are obviously too
large to analyze, avoiding wasted time on checks that would only
time out.