Constant-fold over Op::ITE
If one argument is an ITE instruction where both values are
constants, and the other argument is also a constant, then
the instruction is constant-folded (if possible) into the ITE
instruction. Here, "zext x" of a boolean x is treated as an
ITE instruction "ite x, 1, 0", and "sext x" is treated as
"ite x, -1, 0".