Improve UB handling for Op::IS_NONCANONICAL_NAN
We mark the use of a non-canonical NaN as UB in src to constrain values
in order to prevent spurious failure reports when a non-canonical NaN
has a different bit pattern in tgt. We can remove these if all uses are
in floating-point instructions, as they always canonicalize NaNs, so the
result is the same regardless of whether the input is canonical or not.

It is in general unsafe to remove UB instructions, so we introduce a new
UB_NAN instruction that functions like UB but is used exclusively for
constraining non-canonical NaNs.