Remove Op::UB_NAN
The non-canonical NaN handling is needed because src and tgt may
produce different values for NaN, and we would therefore report
that the program is miscompiled. But the non-canonical NaN checking
is expensive -- especially when checking loaded values. The idea
behind Op::UB_NAN was to allow elimination of these checks when
all uses of the value occur in floating-point instructions (since
they canonicalize the value as a side effect).

But this did not work reliably: there are more cases than I
anticipated where GCC may eliminate floating-point operations.
We could address this by special-casing the conditions that can
be eliminated, but that is brittle and and we are missing
infrastructure for doing this...

Therefore, remove Op::UB_NAN handling for now until we can
implement a better solution.