Do not hard-code bitsize in gen_clz/gen_clrsb
The return value was created with a hard-coded bitsize to generate
identical IR for the code produced from GIMPLE and the RISC-V assembly.
But AArch64 vector operations are often performed on smaller sizes,
so using the correct type from the start avoids unnecessary truncations.

And `gimple_conv.cpp` has been changed, so we were not generating
identical IR anyway.