Improve signbit
signbit(x) returns nonzero if the sign of x is negative, which we
had implemented to always return 1. This worked for smtgcc-tv,
but the backends typically emit signbit(x) as the equivalent of
"x & 0x80000000," causing smtgcc-tv-backend to claim the code was
miscompiled.

We now make the builtin generate an Op::AND when the floating-point
value has a smaller bit size than the resulting type.