Do not eliminate all instructions for UB block
The pass used to remove all instructions for basic blocks that
were always UB. This had the bad side effect of always changing
the result for code entering an always UB block, and therefore
makes smtgcc less useful as it is harder to identify issues that
may generate wrong code as compared to cases that only introduce
a new harmless integer overflow.

We now keep all instructions, even for always UB basic blocks.