Cache results in a Redis database
Add support for caching the results of SMT queries in a Redis
database.

Caching is enabled by setting the environment variable SMTGCC_CACHE
to "redis". The result of each SMT query is stored in the database
using a hash of the IR of the check function as the key.

The database is currently hard-coded to host 127.0.0.1 and port 6379.