Handle C++ zero-sized arrays same as in C
GCC has an extension for treating zero-sized arrays, such as:

  int a[0];

as variable-sized when placed last in a struct. The C++ frontend
generated these as arrays with a max index of ~0 instead as not
setting a max index, causing us to always treat accesses as being
in range instead of verifying that it was within valid memory.