Improve indef handling
We treated most operations that may be UB, such as NEGATE_EXPR,
as UB if the input was indef. The idea was that indef represents
"all values," so it would trigger UB.

But this fails in some cases. Consider the following example:

  void f ()
  {
    long v __attribute__((vector_size(16)));
    (void)((v < v) < v);
  }

When compiled for x86_64 with -O0, the GIMPLE looks like:

  _1 = v_2(D) > { 0, 0 };

This is then transformed by veclower so that each element is
processed as:

  _4 = BIT_FIELD_REF <v_2(D), 64, 0>;
  _5 = _4 > 0;
  _6 = (signed long) _5;
  _7 = -_6;
  _8 = (<signed-boolean:64>) _7;

Here, _7 is indef, but it can only have the values 0 and 1, so it
is incorrect to mark _8 as UB due to invalid input.

This patch removes the special-casing of UB checks for indef. The
indef bits are symbolic values, so the normal code already does
the right thing.