Remove unused local memory instructions
Previously, we cloned missing local memory to ensure the memory
layout was identical between src and tgt. The goal was to speed
up checking by enabling more cases to be CSEd. It did help some
cases, but the memory diverges the first time this memory is
written, so it did not help much. And it often made the result
slower as the SMT solver had more memory to keep track of.

This issue was particularly noticeable with smtgcc-tv-backend.so,
where assembly code often begins by storing a register to the
stack, therefore making it impossible to CSE any subsequent
memory operations.