Throw Not_implemented instead of assert
The code handling GIMPLE_UNARY_RHS for vectors assumes the LHS and
RHS have the same number of vector elements. This assumption is
incorrect for VEC_DUPLICATE_EXPR, which caused a type-checking
assert to fail before throwing a Not_implemented exception for
VEC_DUPLICATE_EXPR.

Add an early check for VEC_DUPLICATE_EXPR for now.