aarch64: simplify is_vreg, is_zreg, etc.
is_vreg, is_zreg, etc., performed partial register validation,such
as checking that the register had one or two digits followed by `.`.
But this was not entirely correct -- for example, a z-register does
not need to have `.` efter the digits when used as:

  ldr     z30, [x1]

These functions are only used to determine the parsing strategy
(e.g., SVE or SIMD) or determine if optional registers are present,
so the validation can be simplified. Full validation is performed
during parsing anyway.