aarch64: improve generation of comparisons
One reason checking AArch64 assembly is slower than checking RISC-V
assembly is that the SMT solver has to do much more work for
comparisons due to how conditions are evaluated using the N, Z, C,
and V flags.

For example, a simple comparison

    void foo(int x, int y)
    {
      if (x < y)
        ...;
    }

is generated as

      %14 = param %3, %5
      %15 = param %4, %5
      br .1

    .1:
      %29 = slt %14, %15
      br %29, .2, .3

in src, but is in tgt generated as

      %14 = param %3, %5
      %15 = param %4, %5
      br .1

    .1:
      %17 = sub %14, %15
      %18 = slt %14, %3
      %19 = slt %15, %3
      %20 = slt %17, %3
      %21 = not %18
      %22 = not %19
      %23 = not %20
      %24 = and %18, %22
      %25 = and %24, %23
      %26 = and %21, %19
      %27 = and %26, %20
      %28 = or %25, %27
      %29 = eq %20, %28
      br %29, .3, .2

While it is possible to simplify the tgt code to match the src
comparison, this approach is fragile since the sequence could
change due to other optimizations. Instead, we introduce pseudo-
registers for the HI, LT, and GT condition codes. These are set
to the expected comparison results for instructions like cmp, or
computed explicitly for other instructions.

Instructions that consume the N, Z, C, and V flags can then use
these pseudo-flags when possible, so tgt is now generated identically
to src.