Improve phi-node conversion
Simplify ITE conditions while building the chain of ITE instructions
for phi-nodes.

If we have a phi-node with three arguments, where the path conditions
for the elements are:
  1. a & b & c
  2. a & b & !c
  3. !a & d
we used to build the chain as:
  ite (!a & d), v3, (ite (a & b & !c), v2, v1)
But we know that if the phi-node value is used, then at least one of
the paths must be taken. So we can eliminate a & b in the last ITE:
  ite (!a & d), v3, (ite (!c), v2, v1)
because if 3. is not true, then one of 1. or 2. must be true.

In general, if all conditions in a sub-chain share the same element,
that element can be eliminated.

This optimization makes it possible to CSE more between src and tgt
for optimizations that modify the CFG.