Perform UB check as first check in check_refine
Update check_refine to perform the UB check before checking
the others. The previous behavior aimed to make it easier to
determine whether an UB failure was harmless or affected the
result. But it could also lead to confusing failures in
smtgcc-tv-backend due to the missing loop canonicalization.

Performing the UB check first also enables us to eliminate
instructions that only affect the result if the program has
already failed the UB check. For example, replacing a load
from constant memory with the constant value directly. This
may make a big improvement for smtgcc-tv-backend, as assembly
often reads constants from memory (since it is cheaper than
creating the constant in a register).