Calculate indef after BIT_AND_EXPR/BIT_IOR_EXPR
Calculate the value of the operation before propagating indef.
This makes the IR nicer to read, as the indef calculation
may (after simplification) be equivalent to the operation,
but in the other De Morgan form, and having it first will make
the indef form be used as the canonical form for CSE:ing
the operation, making the IR rather confusing to read.