Canonicalize commutative operations
It is often the case that the IR generated from a function compiled
with different architecture flags (for example, RISC-V -march=rv64gc
and -march=rv64gcb) is nearly identical, but with some commutative
operations generated with the operands in the other order, and
therefore does not hit the cache.

Solve this by using a consistent argument order when creating
commutative binary operations.