Improve fp to integer type converter UB check
The placeholder implementation UB check for out-of-range values
did not correctly handle values close to being out of range.

The reason is that we checked if a floating-point value f was
in range of a 32-bit signed integer as
  f <= (float)(int)0x7fffffff
but this is rounded up when the floating-point mantissa has fewer
than 32 bits, and it therefore calculates
  f <= (float)(unsigned)0x80000000

There are still some potential rounding issues in this code. It
should be replaced with a better check that is both efficient
and correct.