New pass to reduce bitsize
Reduce the bitsize of instructions when only part of the result is
used. For example, a 32-bit addition
  %3 = add %1, %2
  %4 = extract %3, 15, 0
can be changed to a 16-bit addition by truncating its arguments.

This helps when checking RISC-V and AArch64 assembly, as the
generated code often does operations on the full register width
and then truncates the result. The bitsize reduction done by
this pass allows more code to be CSEd between the IR generated
from GIMPLE and the assembly.