aarch64: Improve fcmp/fccmp
The pseudo condition flags were inverted compared to our IR
canonical form. This did not matter for integer comparisons
because !(x < y) can be changed to x >= y. But this does not
work for floating-point comparisons because of NaN.

Change the flags to the inverted form, and enable that for
floating-point comparison too.