Handle uninitialized arguments in GIMPLE_COND
Our handling of EQ_EXPR and NE_EXPR in GIMPLE_ASSIGN has a special
case for comparing arguments where some bits are uninitialized,
where we treat the result as defined if it always returns the same
value regardless of the value of the uninitialized bits.

Enable this for GIMPLE_COND too.