Improve BIT_FIELD_REF
Handle BIT_FIELD_REF in the same way as ARRAY_REF and COMPONENT_REF,
so that only the extracted bits must be in bounds.

This patch is needed because ifcombine may change
  _5 = BIT_FIELD_REF <display_10(D)->platform.D.54265.D.54263, 8, 32>;
  _6 = _5 & 8;
  if (_6 != 0)
to
  _12 = BIT_FIELD_REF <*display_10(D), 64, 64>;
  _13 = _12 & 36507222016;
  if (_13 != 0)
which failed if display_10 was a large structure that overflowed the
memory object, even though the extracted bits were within bounds.

This also helps cases where we read from gigantic arrays, such as
  struct A { char b; char c[0x21000000]; };
  struct A d;
  int
  foo ()
  {
    return d.c[0x20000000] || d.c[1];
  }
which previously gave up with the message "load size too big".