Improve Op::ITE simplification
The converter traverses the chain of array store instructions to
find redundant stores. This is especially useful for indef, as
most Op::ARRAY_SET_INDEF writes 0 and can usually be eliminated.
The array stores generated for .MASK_LEN_STORE have non-trivial
value that cannot be eliminated, but this is then changed to a
constant by the recent Op::ITE simplifications. But simplify_inst
did not traverse the array to see if the new array store could
be eliminated.

Fix this by moving the array chain traversal to simplify_inst.