Handle Op::ITE in specialize_cond_arg
We skipped Op::ITE in specialize_cond_arg -- the idea was that it
should always have been handled by the usual Op::ITE optimization
of consecutive Op::ITE. But the way we handle store instructions
may in fact give us an Op::ITE here that can be optimized.

Fixing this eliminates ~3% of smtgcc-tv-backend.so timeouts when
compiling the GCC test-suite C and C++ files with -O3 -march=rv64gcbv.