Change assert to if-statement
ult_upper_bound/ult_lower_bound had an assert that verified the
input was in the optimized canonical form, which eliminate some
corner cases that cannot be optimized. But it is sometimes useful
to disable simplify_inst when debugging, so this patch changes
it to an if-statement that skips the instruction instead of
triggering an assert.